# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NewsAggregator Api",
        "version": "1.0.0"
    },
    "paths": {
        "/api/news": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Get a listing of news articles",
                "operationId": "getNews",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Article title"
                                                    },
                                                    "source": {
                                                        "type": "string",
                                                        "example": "News source"
                                                    },
                                                    "category": {
                                                        "type": "string",
                                                        "example": "Article category"
                                                    },
                                                    "desc": {
                                                        "type": "string",
                                                        "example": "Article description"
                                                    },
                                                    "url": {
                                                        "type": "string",
                                                        "example": "https://example.com/article"
                                                    },
                                                    "image": {
                                                        "type": "string",
                                                        "example": "https://example.com/article/image.jpg"
                                                    },
                                                    "date": {
                                                        "type": "string",
                                                        "example": "2023-06-27"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Request failed, could not process your request at the moment. Please try again."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/news/search": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Search for news articles",
                "operationId": "searchNews",
                "parameters": [
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "Keyword to search",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "source",
                        "in": "query",
                        "description": "Source to search",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Date to filter by",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Category to filter by",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Article title"
                                                    },
                                                    "source": {
                                                        "type": "string",
                                                        "example": "News source"
                                                    },
                                                    "category": {
                                                        "type": "string",
                                                        "example": "Article category"
                                                    },
                                                    "desc": {
                                                        "type": "string",
                                                        "example": "Article description"
                                                    },
                                                    "url": {
                                                        "type": "string",
                                                        "example": "https://example.com/article"
                                                    },
                                                    "image": {
                                                        "type": "string",
                                                        "example": "https://example.com/article/image.jpg"
                                                    },
                                                    "date": {
                                                        "type": "string",
                                                        "example": "2023-06-27"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Request failed, could not process your request at the moment. Please try again."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}